Moving the score reporting to the lasngfuse module for better reusability
