+++ aieng-eval-agents/aieng/agent_evals/langfuse.py
@@ -11,11 +11,15 @@ from typing import Any, Literal
 from aieng.agent_evals.async_client_manager import AsyncClientManager
 from aieng.agent_evals.configs import Configs
 from aieng.agent_evals.progress import track_with_progress
+from langfuse import Langfuse
+from langfuse.api.resources.commons.types.observations_view import ObservationsView
+from langfuse.api.resources.observations.types.observations_views import ObservationsViews
 from opentelemetry import trace
 from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchSpanProcessor, SimpleSpanProcessor
+from tenacity import retry, stop_after_attempt, wait_exponential
 
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s %(levelname)s %(name)s: %(message)s")
@@ -370,3 +374,122 @@ def _normalize_dataset_record(item: Any, record_number: int) -> dict[str, Any]:
         "expected_output": item["expected_output"],
         "metadata": metadata,
     }
+
+
+def report_usage_scores(
+    trace_id: str,
+    token_threshold: int = 0,
+    latency_threshold: int = 0,
+    cost_threshold: float = 0,
+) -> None:
+    """Report usage scores to Langfuse for a given trace ID.
+
+    WARNING: Due to the nature of the Langfuse API, this function may hang
+    while trying to fetch the observations.
+
+    Parameters
+    ----------
+    trace_id: str
+        The ID of the trace to report the usage scores for.
+    token_threshold: int
+        The token threshold to report the score for.
+        if the token count is greater than the threshold, the score
+        will be reported as 0.
+        Optional, default to 0 (no reporting).
+    latency_threshold: int
+        The latency threshold in seconds to report the score for.
+        if the latency is greater than the threshold, the score
+        will be reported as 0.
+        Optional, default to 0 (no reporting).
+    cost_threshold: float
+        The cost threshold to report the score for.
+        if the cost is greater than the threshold, the score
+        will be reported as 0.
+        Optional, default to 0 (no reporting).
+    """
+    langfuse_client = AsyncClientManager.get_instance().langfuse_client
+    observations = _get_observations_with_retry(trace_id, langfuse_client)
+
+    if token_threshold > 0:
+        total_tokens = sum(_obs_attr(observation, "totalTokens") for observation in observations.data)
+        if total_tokens <= token_threshold:
+            score = 1
+            comment = "Token count is less than or equal to the threshold."
+        else:
+            score = 0
+            comment = "Token count is greater than the threshold."
+
+        logger.info("Reporting score for token count")
+        langfuse_client.create_score(
+            name="Token Count",
+            value=score,
+            trace_id=trace_id,
+            comment=comment,
+            metadata={
+                "total_tokens": total_tokens,
+                "token_threshold": token_threshold,
+            },
+        )
+
+    if latency_threshold > 0:
+        total_latency = sum(_obs_attr(observation, "latency") for observation in observations.data)
+        if total_latency <= latency_threshold:
+            score = 1
+            comment = "Latency is less than or equal to the threshold."
+        else:
+            score = 0
+            comment = "Latency is greater than the threshold."
+
+        logger.info("Reporting score for latency")
+        langfuse_client.create_score(
+            name="Latency",
+            value=score,
+            trace_id=trace_id,
+            comment=comment,
+            metadata={
+                "total_latency": total_latency,
+                "latency_threshold": latency_threshold,
+            },
+        )
+
+    if cost_threshold > 0:
+        total_cost = sum(_obs_attr(observation, "calculated_total_cost") for observation in observations.data)
+        if total_cost <= cost_threshold:
+            score = 1
+            comment = "Cost is less than or equal to the threshold."
+        else:
+            score = 0
+            comment = "Cost is greater than the threshold."
+
+        logger.info("Reporting score for cost")
+        langfuse_client.create_score(
+            name="Cost",
+            value=score,
+            trace_id=trace_id,
+            comment=comment,
+            metadata={
+                "total_cost": total_cost,
+                "cost_threshold": cost_threshold,
+            },
+        )
+
+    langfuse_client.flush()
+
+
+def _obs_attr(observation: ObservationsView, attribute: str) -> Any:
+    """Get the value of an attribute from an observation."""
+    attribute_value = getattr(observation, attribute)
+    if attribute_value == 0:
+        logger.error(f"Observation attribute value for {attribute} is 0")
+        return 0
+    if attribute_value is None:
+        logger.error(f"Observation attribute value for {attribute} is None")
+        return 0
+    return attribute_value
+
+
+@retry(stop=stop_after_attempt(10), wait=wait_exponential(multiplier=1, min=5, max=15))
+def _get_observations_with_retry(trace_id: str, langfuse_client: Langfuse) -> ObservationsViews:
+    """Get the observations for a given trace ID with retry/backoff."""
+    logger.info(f"Getting observations for trace {trace_id}...")
+    return langfuse_client.api.observations.get_many(trace_id=trace_id, type="GENERATION")
